+++ app/src/main/kotlin/com/aliucord/manager/installer/steps/patch/ReplaceIconStep.kt
@@ -16,18 +16,18 @@ import com.aliucord.manager.installer.util.ArscUtil.getResourceFileName
 import com.aliucord.manager.installer.util.AxmlUtil
 import com.aliucord.manager.ui.screens.installopts.InstallOptions
 import com.aliucord.manager.ui.screens.installopts.InstallOptions.IconReplacement
-import com.aliucord.manager.util.find
 import com.aliucord.manager.util.getResBytes
-import com.github.diamondminer88.zip.ZipReader
 import com.github.diamondminer88.zip.ZipWriter
-import com.google.devrel.gmscore.tools.apk.arsc.*
+import com.google.devrel.gmscore.tools.apk.arsc.BinaryResourceValue
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
-import java.io.File
 import java.io.InputStream
 
 /**
- * Replace icons
+ * Patch the mipmap-v26 launcher icons' background, foreground and monochrome attributes,
+ * and if below API 26, replace the traditional mipmap icons instead.
+ * All filenames are retrieved based on the resource definitions in `resources.arsc`,
+ * so that this works on pretty much almost any APK.
  */
 @Stable
 class ReplaceIconStep(private val options: InstallOptions) : Step(), KoinComponent {
@@ -45,7 +45,7 @@ class ReplaceIconStep(private val options: InstallOptions) : Step(), KoinCompone
         val apk = container.getStep<CopyDependenciesStep>().patchedApk
         val arsc = ArscUtil.readArsc(apk)
 
-        val iconRscIds = readManifestIconInfo(apk)
+        val iconRscIds = AxmlUtil.readManifestIconInfo(apk)
         val monochromeRscId = if (!options.monochromeIcon) null else {
             val filePathIdx = arsc.getMainArscChunk().stringPool
                 .addString("res/ic_aliucord_monochrome.xml")
@@ -58,6 +58,9 @@ class ReplaceIconStep(private val options: InstallOptions) : Step(), KoinCompone
                 valueData = filePathIdx,
             )
         }
+
+        val typeChunks = arsc.getPackageChunk().getTypeChunks("mipmap")
+        println("type: mipmap, configs: ${typeChunks.map { it.configuration.toString() }}")
 
         val squareIconFile = arsc.getMainArscChunk().getResourceFileName(iconRscIds.squareIcon, "anydpi-v26")
         val roundIconFile = arsc.getMainArscChunk().getResourceFileName(iconRscIds.roundIcon, "anydpi-v26")
@@ -87,9 +90,6 @@ class ReplaceIconStep(private val options: InstallOptions) : Step(), KoinCompone
         ZipWriter(apk, /* append = */ true).use {
             // val squareIcon = readAsset("icons/ic_logo_square.png")
             // val roundIcon = readAsset("icons/ic_logo_round.png")
-            // val replacements = mapOf(
-            //     arrayOf("_h_.png", "9MB.png", "Dy7.png", "kC0.png", "oEH.png", "RG0.png", "ud_.png", "W_3.png") to squareIcon
-            // )
             //
             // for ((files, replacement) in replacements) {
             //     for (file in files) {
@@ -107,47 +107,6 @@ class ReplaceIconStep(private val options: InstallOptions) : Step(), KoinCompone
             it.writeEntry("resources.arsc", newArscBytes)
         }
     }
-
-    /**
-     * From an APK, read the manifest's `icon` and `roundIcon` references to a resource.
-     * This is then used to get the filename of the resource from `resources.arsc`.
-     */
-    private fun readManifestIconInfo(apk: File): ManifestIconInfo {
-        val manifestBytes = ZipReader(apk).use { it.openEntry("AndroidManifest.xml")!!.read() }
-        val manifest = BinaryResourceFile(manifestBytes)
-        val mainChunk = manifest.chunks.single() as XmlChunk
-
-        // Prefetch string indexes to avoid parsing the entire string pool
-        val iconStringIdx = mainChunk.stringPool.indexOf("icon")
-        val roundIconStringIdx = mainChunk.stringPool.indexOf("roundIcon")
-        val applicationStringIdx = mainChunk.stringPool.indexOf("application")
-
-        val applicationChunk = mainChunk.chunks
-            .find { it is XmlStartElementChunk && it.nameIndex == applicationStringIdx } as? XmlStartElementChunk
-            ?: error("Unable to find <application> in manifest")
-
-        val squareIcon = applicationChunk.attributes
-            .find { it.nameIndex() == iconStringIdx }
-            ?: error("Unable to find android:icon in manifest")
-
-        val roundIcon = applicationChunk.attributes
-            .find { it.nameIndex() == roundIconStringIdx }
-            ?: error("Unable to find android:roundIcon in manifest")
-
-        assert(squareIcon.typedValue().type() == BinaryResourceValue.Type.REFERENCE)
-        assert(roundIcon.typedValue().type() == BinaryResourceValue.Type.REFERENCE)
-
-        return ManifestIconInfo(
-            // Resource IDs into resources.arsc
-            squareIcon = BinaryResourceIdentifier.create(squareIcon.typedValue().data()),
-            roundIcon = BinaryResourceIdentifier.create(roundIcon.typedValue().data()),
-        )
-    }
-
-    private data class ManifestIconInfo(
-        val squareIcon: BinaryResourceIdentifier,
-        val roundIcon: BinaryResourceIdentifier,
-    )
 
     private fun readAsset(fileName: String): ByteArray =
         context.assets.open(fileName).use(InputStream::readBytes)
